important modifications (see below)
in the main function fishmap.R I had to explicitly import ggplot2,
otherwise we would have to bind ggplot2:: to all functions.
The data() call needs to be revised yet. We must put this database
inside the package and not call it from marelac.
Still getting note of invisible bindings from some objects.